+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.7.0
+
+* **Breaking change:** removed deprecated static methods in `FeatureDiscovery`.
+* **Breaking change:** removed deprecated parameters in the `EnsureVisible` constructor.
+* **Breaking change:** overlays will always be dismissed when calling `FeatureDiscovery.dismissAll`.
+* *Deprecated* `activeFeatureId`; replaced by `currentFeatureIdOf` to emphasize that this is a getter.
+* *Deprecated* `FeatureDiscovery.dismiss`; replaced by `dismissAll` to indicate that no next step
+  will be shown.
+* Added `assert` to require at least one step to be passed to `FeatureDiscovery.discoverFeatures`.
+* Incorrect documentation of some static methods in `FeatureDiscovery` has been updated.
+* Error messages have been improved : the error thrown when the widget tree isn't wrapped in
+  a `FeatureDiscovery` widget is clearer.
+* Incorrect behavior when `onDismiss` returned `Future<false>` has been fixed.
+
 ## 0.6.1
 
 * Update version constraint to `^4.0.1` for `provider` dependency.